add checks for docker

+++ Scripts/run-tests-linux-docker.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+command -v docker > /dev/null 2&>1 || { echo >&2 "Install docker https://www.docker.com"; exit 1; }
+
 IMAGE=pathostesting
-docker image rm -f "$IMAGE" 2>/dev/null
+docker image rm -f "$IMAGE" 2&>1 > /dev/null
 docker build -t "$IMAGE" -f Scripts/Dockerfile-testing-linux . && docker run --rm "$IMAGE"
